uint32_t used

+++ shm/cargo.cpp
@@ -1,11 +1,11 @@
 #include "cargo.hpp"
 
-Cargo& Cargo::operator+=(uint8_t amount) {
+Cargo& Cargo::operator+=(uint32_t amount) {
     amount_ += amount;
     return *this;
 }
 
-Cargo& Cargo::operator-=(uint8_t amount) {
+Cargo& Cargo::operator-=(uint32_t amount) {
     amount_ -= amount;
     return *this;
 }
